feat(go.d/snmp): add SNMP sysObjectID mappings for device identification (#20805)
(cherry picked from commit ec25d71fab168db49c4b65a338efccf4afde1d18)
