+++ src/go/plugin/go.d/collector/snmp/collect.go
@@ -150,11 +150,18 @@ func (c *Collector) setupVnode(si *snmpsd.SysInfo, deviceMeta map[string]map[str
 	if _, ok := labels["location"]; !ok && si.Location != "" {
 		labels["location"] = si.Location
 	}
-	if _, ok := labels["vendor"]; !ok && si.Organization != "" {
-		labels["vendor"] = si.Organization
-		if v, ok := orgToVendorMap[si.Organization]; ok {
-			labels["vendor"] = v
+	if _, ok := labels["vendor"]; !ok {
+		if si.Vendor != "" {
+			labels["vendor"] = si.Vendor
+		} else if si.Organization != "" {
+			labels["vendor"] = si.Organization
 		}
+	}
+	if _, ok := labels["type"]; !ok && si.Category != "" {
+		labels["type"] = si.Category
+	}
+	if _, ok := labels["model"]; !ok && si.Model != "" {
+		labels["model"] = si.Model
 	}
 
 	return &vnodes.VirtualNode{
@@ -189,133 +196,6 @@ func pduToInt(pdu gosnmp.SnmpPDU) (int64, error) {
 	default:
 		return 0, fmt.Errorf("unussported type: '%v'", pdu.Type)
 	}
-}
-
-var orgToVendorMap = map[string]string{
-	"Alcatel-Lucent TMC (formerly 'Alcatel SOC')":      "Alcatel-Lucent",
-	"allied networks GmbH":                             "Allied",
-	"Allied Data Technologies":                         "Allied",
-	"Allied Telesis, Inc.":                             "Allied",
-	"American Power Conversion Corp.":                  "APC",
-	"Arista Networks, Inc. (formerly 'Arastra, Inc.')": "Arista",
-	"Aruba PEC S.p.A.":                                 "Aruba",
-	"Aruba S.r.l.":                                     "Aruba",
-	"Aruba, a Hewlett Packard Enterprise company":      "Aruba",
-	"AVAYA":               "Avaya",
-	"Avaya Atlanta Lab":   "Avaya",
-	"Avaya Communication": "Avaya",
-	"Barracuda Networks AG (previous was 'phion Information Technologies')": "Barracuda",
-	"Barracuda Networks, Inc.":                                                "Barracuda",
-	"barracuda digitale agentur GmbH":                                         "Barracuda",
-	"Blade Network Technologies, Inc.":                                        "IBM",
-	"Brocade Communication Systems, Inc. (formerly 'Foundry Networks, Inc.')": "Brocade",
-	"Brocade Communication Systems, Inc. (formerly 'Rhapsody Networks Inc.')": "Brocade",
-	"Brocade Communications Systems, Inc.":                                    "Brocade",
-	"Brocade Communications Systems, Inc. (formerly 'McDATA Corp.')":          "Brocade",
-	"Brocade Communications Systems, Inc. (formerly 'McDATA Corporation')":    "Brocade",
-	"Brocade Communications Systems, Inc. (formerly 'McDATA,Inc')":            "Brocade",
-	"Brocade Communications Systems, Inc. (formerly 'NuView Inc.')":           "Brocade",
-	"CIENA Corporation (formerly 'ONI Systems Corp.')":                        "Ciena",
-	"Ciena (formerly 'Akara Inc.')":                                           "Ciena",
-	"Ciena Corporation":                                                       "Ciena",
-	"Ciena Corporation (formerly 'Catena Networks')":                          "Ciena",
-	"Cisco Flex Platform":                                                     "Cisco",
-	"Cisco Sera":                                                              "Cisco",
-	"Cisco SolutionsLab":                                                      "Cisco",
-	"Cisco Sytems, Inc.":                                                      "Cisco",
-	"Cisco Systems":                                                           "Cisco",
-	"Cisco Systems Inc":                                                       "Cisco",
-	"Cisco Systems India Private Limited":                                     "Cisco",
-	"Cisco Systems, Inc.":                                                     "Cisco",
-	"Cisco Systems, Inc. (formerly 'Arch Rock Corporation')":                  "Cisco",
-	"ciscoSystems":                                                            "Cisco",
-	"Citrix Systems Inc.":                                                     "Citrix",
-	"Dialogic Corporation":                                                    "Dialogic",
-	"D-Link Systems, Inc.":                                                    "D-Link",
-	"Dell Inc.":                                                               "Dell",
-	"Eaton Energy Automation Solutions (EAS) Division":                        "Eaton",
-	"EATON Wireless":                                                          "Eaton",
-	"Ericsson AB":                                                             "Ericsson",
-	"Ericsson AB - 4G5G (formerly 'Ellemtel Telecommunication Systems Laboratories')": "Ericsson",
-	"Ericsson AB - Packet Core Networks":                                              "Ericsson",
-	"Ericsson Ahead Communications Systems GmbH":                                      "Ericsson",
-	"Ericsson Communications Ltd.":                                                    "Ericsson",
-	"Ericsson Denmark A/S, Telebit Division":                                          "Ericsson",
-	"Ericsson Inc. (formerly 'BelAir Networks')":                                      "Ericsson",
-	"Ericsson Mobile Platforms AB":                                                    "Ericsson",
-	"Ericsson Nikola Tesla d.d.":                                                      "Ericsson",
-	"Ericsson Research Montreal (LMC)":                                                "Ericsson",
-	"Ericsson Wireless LAN Systems":                                                   "Ericsson",
-	"ERICSSON FIBER ACCESS":                                                           "Ericsson",
-	"Extreme Networks":                                                                "Extreme",
-	"Extreme Networks (formerly 'Ipanema Technologies')":                              "Extreme",
-	"F5 Labs, Inc.":                                                                   "F5",
-	"F5 Networks Inc":                                                                 "F5",
-	"Fortinet, Inc.":                                                                  "Fortinet",
-	"Fortinet. Inc.":                                                                  "Fortinet",
-	"Gigamon Systems LLC":                                                             "Gigamon",
-	"Hewlett-Packard":                                                                 "HP",
-	"Hewlett-Packard (Schweiz) GmbH":                                                  "HP",
-	"Hewlett-Packard Slovakia":                                                        "HP",
-	"Hewlett Packard Enterprise":                                                      "HPE",
-	"HUAWEI Technology Co.,Ltd":                                                       "Huawei",
-	"Huawei Symantec Technologies Co.,Ltd":                                            "Huawei",
-	"Infinera Corp.":                                                                  "Infinera",
-	"InfoBlox Inc.":                                                                   "Infoblox",
-	"Infoblox, WinConnect (formerly 'Ipanto')":                                        "Infoblox",
-	"Juniper Financial Corp.":                                                         "Juniper",
-	"Juniper Networks, Inc.":                                                          "Juniper",
-	"Juniper Networks/Funk Software":                                                  "Juniper",
-	"Juniper Networks/Unisphere":                                                      "Juniper",
-	"KYOCERA Corporation":                                                             "Kyocera",
-	"Kyocera Communication Systems Co.Ltd":                                            "Kyocera",
-	"McAFee Associates Inc.":                                                          "McAfee",
-	"McAfee (formerly 'Secure Computing Corporation')":                                "McAfee",
-	"McAfee Inc. (formerly 'Network Associates, Inc.')":                               "McAfee",
-	"McAfee, Inc.  (formerly 'Securify, Inc.')":                                       "McAfee",
-	"McAfee Inc. (formerly 'Reconnex Corporation')":                                   "McAfee",
-	"McAfee, LLC":                                                                     "McAfee",
-	"Meraki Networks, Inc.":                                                           "Meraki",
-	"Nasuni Corporation":                                                              "Nasuni",
-	"NEC Corporation":                                                                 "NEC",
-	"NEC Eluminant Technologies, Inc.":                                                "NEC",
-	"NEC Platforms, Ltd.":                                                             "NEC",
-	"NEC Telenetworx,Ltd":                                                             "NEC",
-	"NEC COMPUTERS INTERNATIONAL B.V.":                                                "NEC",
-	"NEC informatec systems,ltd.":                                                     "NEC",
-	"NEC Electronics Corporation":                                                     "NEC",
-	"NEC Unified Solutions":                                                           "NEC",
-	"NEC Enterprise Communication Technologies":                                       "NEC",
-	"Network Appliance Corporation":                                                   "NetApp",
-	"Nokia (formerly 'Alcatel-Lucent')":                                               "Nokia",
-	"Nokia (formerly 'Novarra, Inc.')":                                                "Nokia",
-	"Nokia Distributed Access":                                                        "Nokia",
-	"Nokia Networks (formerly 'Nokia Siemens Networks')":                              "Nokia",
-	"Nokia Shanghai Bell":                                                             "Nokia",
-	"NVIDIA Corporation":                                                              "NVIDIA",
-	"PALO ALTO NETWORKS":                                                              "Palo Alto",
-	"Palo Alto Research Center, Inc.":                                                 "Palo Alto",
-	"Palo Alto Software, Inc.":                                                        "Palo Alto",
-	"Ruckus Wireless, Inc.":                                                           "Ruckus",
-	"SINETICA":                                                                        "Panduit",
-	"Sophos Plc":                                                                      "Sophos",
-	"SVTO Hewlett-Packard":                                                            "HP",
-	"Synology Inc.":                                                                   "Synology",
-	"Tejas Networks":                                                                  "Tejas",
-	"TP-Link Systems Inc.":                                                            "TP-Link",
-	"Ubiquiti Networks, Inc.":                                                         "Ubiquiti",
-	"Velocloud Networks, Inc.":                                                        "VeloCloud",
-	"Vertiv (formerly 'Emerson Computer Power')":                                      "Vertiv",
-	"Vertiv (formerly 'Emerson Energy Systems')":                                      "Vertiv",
-	"Vertiv Tech Co.,Ltd. (formerly 'Emerson Network Power Co.,Ltd.')":                "Vertiv",
-	"Vertiv (formerly 'Geist Manufacturing, Inc')":                                    "Vertiv",
-	"Vertiv Co":                                                                       "Vertiv",
-	"VMware Inc.":                                                                     "VMware",
-	"WatchGuard Technologies Inc.":                                                    "WatchGuard",
-	"Western Digital Corporation":                                                     "Western Digital",
-	"Yokogawa-Hewlett-Packard":                                                        "HP",
-	"Zebra Technologies Corporation":                                                  "Zebra",
-	"ZyXEL Communications Corp.":                                                      "Zyxel",
 }
 
 //func physAddressToString(pdu gosnmp.SnmpPDU) (string, error) {
